+++ src/utils.c
@@ -22,8 +22,8 @@ static int operations_priority(char first_opr, char second_opr)
 
 void whitespace_cleaner(char* str)
 {
-    size_t j = 0;
-    for (size_t i = 0; str[i] != '\0'; i++) {
+    int j = 0;
+    for (int i = 0; str[i] != '\0'; i++) {
         if (str[i] != ' ') {
             str[j++] = str[i];
         }
@@ -35,7 +35,7 @@ bool is_double(char* str)
 {
     if (strlen(str) == 1 && str[0] == '-')
         return false;
-    for (size_t i = 0; str[i] != '\0'; i++) {
+    for (int i = 0; str[i] != '\0'; i++) {
         if (!isdigit(str[i]) && str[i] != '-' && str[i] != '.')
             return false;
     }
@@ -55,13 +55,13 @@ TokenArray to_rpn(TokenArray infix_expr)
         .size = 0
     };
 
-    for (size_t i = 0; i < MAX_LENGTH; i++) {
+    for (int i = 0; i < MAX_LENGTH; i++) {
         postfix_expr.array[i] = calloc(MAX_LENGTH, sizeof(char*));
     }
 
-    size_t j = 0;
+    int j = 0;
 
-    for (size_t i = 0; i < infix_expr.size; i++) {
+    for (int i = 0; i < infix_expr.size; i++) {
         if (is_double(infix_expr.array[i]) || is_function(infix_expr.array[i])) {
             strcpy(postfix_expr.array[j++], infix_expr.array[i]);
             continue;
